+++ web/source/components/editor-controls.jsx
@@ -24,9 +24,9 @@ var Component = React.createClass({
 					{layoutElements}
 				</div>
 				<div className="btn-group" role="group">
-					<button onClick={actions.addLayout} type="button" className="btn btn-success">Add Layout</button>
-					<button onClick={this.remove} type="button" className="btn btn-danger">Remove Layout {this.state.currentLayout + 1}</button>
-					<a href="#/settings/engine" type="button" className="btn btn-default" style={{textShadow: 'none'}}>Change Keyboard Layout</a>
+					<button onClick={actions.addLayout} type="button" className="btn btn-success"><i className="glyphicon glyphicon-plus"/> Layout</button>
+					<button onClick={this.remove} type="button" className="btn btn-danger"><i className="glyphicon glyphicon-trash"/> Layout {this.state.currentLayout + 1}</button>
+					<a href="#/settings/engine" type="button" className="btn btn-default" style={{textShadow: 'none'}}><i className="glyphicon glyphicon-transfer"/> Change Keyboard Layout</a>
 				</div>
 			</div>
 		);